# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_group_role_assignments_create_groups__group_id__role_assignments_post": {
        "properties": {
          "contract_signed": {
            "default": false,
            "title": "Contract Signed",
            "type": "boolean"
          },
          "role_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          },
          "term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          },
          "volunteer_ids": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Volunteer Ids",
            "type": "array"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "title": "Body_group_role_assignments_create_groups__group_id__role_assignments_post",
        "type": "object"
      },
      "Body_group_roles_create_groups__group_id__roles_post": {
        "properties": {
          "pingvin_points": {
            "title": "Pingvin Points",
            "type": "integer"
          },
          "role_name": {
            "title": "Role Name",
            "type": "string"
          }
        },
        "required": [
          "role_name",
          "pingvin_points"
        ],
        "title": "Body_group_roles_create_groups__group_id__roles_post",
        "type": "object"
      },
      "Body_group_roles_update_groups__group_id__roles__role_id__patch": {
        "properties": {
          "pingvin_points": {
            "title": "Pingvin Points",
            "type": "integer"
          },
          "role_name": {
            "title": "Role Name",
            "type": "string"
          }
        },
        "required": [
          "role_name",
          "pingvin_points"
        ],
        "title": "Body_group_roles_update_groups__group_id__roles__role_id__patch",
        "type": "object"
      },
      "Body_groups_apply_semester_transfer_groups__group_id__semester_transfer_post": {
        "properties": {
          "role_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Role Ids",
            "type": "array"
          },
          "target_semester": {
            "title": "Target Semester",
            "type": "integer"
          },
          "volunteer_ids": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Volunteer Ids",
            "type": "array"
          }
        },
        "required": [
          "target_semester"
        ],
        "title": "Body_groups_apply_semester_transfer_groups__group_id__semester_transfer_post",
        "type": "object"
      },
      "Body_groups_create_groups_post": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "active_until_semester": {
            "title": "Active Until Semester",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discount_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Step"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Group Id"
          }
        },
        "required": [
          "name",
          "active_until_semester"
        ],
        "title": "Body_groups_create_groups_post",
        "type": "object"
      },
      "Body_groups_update_groups__group_id__patch": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "active_until_semester": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Until Semester"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "discount_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Step"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Group Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "Body_groups_update_groups__group_id__patch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kvarteret Personal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/groups": {
      "get": {
        "operationId": "groups_index_groups_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Index"
      },
      "post": {
        "operationId": "groups_create_groups_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_groups_create_groups_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Create"
      }
    },
...
    "/groups/{group_id}": {
      "delete": {
        "operationId": "groups_delete_groups__group_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Delete"
      },
      "get": {
        "operationId": "groups_detail_groups__group_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Detail"
      },
      "patch": {
        "operationId": "groups_update_groups__group_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_groups_update_groups__group_id__patch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Update"
      }
    },
    "/groups/{group_id}/archive": {
      "post": {
        "operationId": "groups_archive_groups__group_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Archive"
      }
    },
...
    "/groups/{group_id}/history/{history_id}": {
      "delete": {
        "operationId": "group_history_delete_groups__group_id__history__history_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "history_id",
            "required": true,
            "schema": {
              "title": "History Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Group History Delete"
      }
    },
    "/groups/{group_id}/role-assignments": {
      "post": {
        "operationId": "group_role_assignments_create_groups__group_id__role_assignments_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_group_role_assignments_create_groups__group_id__role_assignments_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Group Role Assignments Create"
      }
    },
    "/groups/{group_id}/roles": {
      "post": {
        "operationId": "group_roles_create_groups__group_id__roles_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_group_roles_create_groups__group_id__roles_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Group Roles Create"
      }
    },
    "/groups/{group_id}/roles/{role_id}": {
      "delete": {
        "operationId": "group_roles_delete_groups__group_id__roles__role_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "title": "Role Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Group Roles Delete"
      },
      "patch": {
        "operationId": "group_roles_update_groups__group_id__roles__role_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "title": "Role Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_group_roles_update_groups__group_id__roles__role_id__patch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Group Roles Update"
      }
    },
    "/groups/{group_id}/semester-transfer": {
      "get": {
        "operationId": "groups_semester_transfer_groups__group_id__semester_transfer_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source_semester",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Semester"
            }
          },
          {
            "in": "query",
            "name": "target_semester",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Semester"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Semester Transfer"
      },
      "post": {
        "operationId": "groups_apply_semester_transfer_groups__group_id__semester_transfer_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_groups_apply_semester_transfer_groups__group_id__semester_transfer_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Groups Apply Semester Transfer"
      }
    },
...
  }
}